updated to size_t

+++ libft.h
@@ -6,13 +6,14 @@
 /*   By: fschmid <fschmid@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/11 09:17:52 by fschmid           #+#    #+#             */
-/*   Updated: 2022/10/11 11:56:39 by fschmid          ###   ########.fr       */
+/*   Updated: 2022/10/11 12:05:35 by fschmid          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 # include <stdbool.h>
+# include <stdlib.h>
 
 bool	ft_isalpha(int c);
 
@@ -26,9 +27,9 @@ bool	ft_isprint(int c);
 
 int		ft_strlen(char *str);
 
-void	*ft_memset(void *b, int c, int len);
+void	*ft_memset(void *b, int c, size_t len);
 
-void	ft_bzero(void *s, int n);
+void	ft_bzero(void *s, size_t n);
 
-void	*ft_memmove(void *dst, const void *src, int len);
+void	*ft_memmove(void *dst, const void *src, size_t len);
 #endif
